ml-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.3
prophet==1.1.6
pillow==11.1.0
pytesseract==0.3.13
openai==1.65.4
python-multipart==0.0.20
openpyxl==3.1.5
joblib==1.4.2
httpx==0.28.1
python-dotenv==1.0.1
